+++ src/ska_low_mccs_spshw/station/tests/base_daq_test.py
@@ -341,6 +341,8 @@ class BaseDaqTest(TpmSelfCheckTest):
         self._stop_directory_watch()
         if self.daq_proxy is not None:
             self.daq_proxy.Stop()
+            # Sleep to allow time to stop.
+            time.sleep(1)
         # self._delete_data()
 
     @contextmanager