+++ src/ska_low_mccs_spshw/tile/tile_simulator.py
@@ -202,6 +202,7 @@ class StationBeamformer:
         self._start_frame = 0
         self._last_frame = 0
         self.station_beam_flag = False
+        self._group_is_running = [False] * 48
 
     def define_channel_table(self: StationBeamformer, table: list[list[int]]) -> None:
         """
@@ -260,17 +261,52 @@ class StationBeamformer:
 
     def start(
         self: StationBeamformer,
+        channel_groups: Optional[list[int]],
     ) -> None:
-        """Start."""
+        """
+        Start.
+
+        :param channel_groups: Groups of channels to start
+        """
+        if channel_groups is None:
+            channel_groups = list(range(48))
         self._is_running = True
+        for group in channel_groups:
+            self._group_is_running[group] = True
 
-    def stop(self: StationBeamformer) -> None:
-        """stop."""
-        self._is_running = False
+    def stop(
+        self: StationBeamformer,
+        channel_groups: Optional[list[int]],
+    ) -> None:
+        """
+        Stop.
 
-    def is_running(self: StationBeamformer) -> bool:
-        """:return: is running."""
-        return self._is_running
+        :param channel_groups: Groups of channels to start
+        """
+        if channel_groups is None:
+            channel_groups = list(range(48))
+        for group in channel_groups:
+            self._group_is_running[group] = False
+        if not any(self._group_is_running):
+            self._is_running = False
+
+    def is_running(
+        self: StationBeamformer,
+        channel_groups: Optional[list[int]] = None,
+    ) -> bool:
+        """
+        Check if the beamformer is running for the specified channels.
+
+        :param channel_groups: Groups of channels to start
+
+        :return: is running.
+        """
+        if channel_groups is None:
+            return self._is_running
+        running = False
+        for group in channel_groups:
+            running = running or self._group_is_running[group]
+        return self._is_running and running
 
     def enable_flagging(self) -> None:
         """Enable station beam flagging."""
@@ -2186,7 +2222,7 @@ class TileSimulator:
         duration: int = -1,
         scan_id: int = 0,
         mask: int | None = None,
-        beam: int | None = None,
+        subarray_beam: int | None = None,
         channel_groups: list[int] | None = None,
     ) -> bool:
         """
@@ -2197,16 +2233,17 @@ class TileSimulator:
         :param scan_id: ID of the scan, to be specified in the CSP SPEAD header
         :param mask: Bitmask of the channels to be started.
             Ignored if beam is specified.
-        :param beam: Beam number to start. Computes the mask using beam table
+        :param subarray_beam: subarray_beam number to start.
+            Computes the mask using beam table
         :param channel_groups: list of channel groups, in range 0:48.
             group 0 for channels 0-7, to group 47 for channels 380-383.
 
         :return: true if the beamformer was started successfully.
         """
-        if self.beamformer_is_running():
+        if self.beamformer_is_running(channel_groups=channel_groups):
             return False
-        self.tpm.beam1.start()  # type: ignore
-        self.tpm.beam2.start()  # type: ignore
+        self.tpm.beam1.start(channel_groups)  # type: ignore
+        self.tpm.beam2.start(channel_groups)  # type: ignore
         return True
 
     @check_mocked_overheating
@@ -2214,7 +2251,7 @@ class TileSimulator:
     def stop_beamformer(
         self: TileSimulator,
         mask: bool | None = None,
-        beam: int | None = None,
+        subarray_beam: int | None = None,
         channel_groups: list[int] | None = None,
     ) -> None:
         """
@@ -2222,12 +2259,13 @@ class TileSimulator:
 
         :param mask: Bitmask of the channels to be started.
             Ignored if beam is specified.
-        :param beam: Beam number to start. Computes the mask using beam table
+        :param subarray_beam: Subarray beam number to start.
+            Computes the mask using beam table
         :param channel_groups: list of channel groups, in range 0:48.
             group 0 for channels 0-7, to group 47 for channels 380-383.
         """
-        self.tpm.beam1.stop()  # type: ignore
-        self.tpm.beam2.stop()  # type: ignore
+        self.tpm.beam1.stop(channel_groups)  # type: ignore
+        self.tpm.beam2.stop(channel_groups)  # type: ignore
 
     @check_mocked_overheating
     @connected
@@ -2695,8 +2733,8 @@ class TileSimulator:
     @connected
     def beamformer_is_running(
         self: TileSimulator,
-        mask: bool | None = None,
-        beam: int | None = None,
+        mask: int | None = None,
+        subarray_beam: int | None = None,
         channel_groups: list[int] | None = None,
     ) -> bool:
         """
@@ -2704,15 +2742,17 @@ class TileSimulator:
 
         :param mask: Bitmask of the channels to be started.
             Ignored if beam is specified.
-        :param beam: Beam number to start. Computes the mask using beam table
+        :param subarray_beam: subarray beam number to start.
+            Computes the mask using beam table
         :param channel_groups: list of channel groups, in range 0:48.
             group 0 for channels 0-7, to group 47 for channels 380-383.
 
         :return: is the beam is running
         """
-        return (
-            self.tpm.beam1.is_running()  # type: ignore
-            and self.tpm.beam2.is_running()  # type: ignore
+        return self.tpm.beam1.is_running(  # type: ignore
+            channel_groups
+        ) and self.tpm.beam2.is_running(  # type: ignore
+            channel_groups
         )
 
     def set_tpm_temperature_thresholds(